refactor: align collection type with provisioning flow (#52)

+++ src/main/java/be/cytomine/appengine/handlers/storage/impl/FileSystemStorageHandler.java
@@ -2,8 +2,10 @@ package be.cytomine.appengine.handlers.storage.impl;
 
 import be.cytomine.appengine.dto.handlers.filestorage.Storage;
 import be.cytomine.appengine.exceptions.FileStorageException;
-import be.cytomine.appengine.handlers.FileData;
-import be.cytomine.appengine.handlers.FileStorageHandler;
+import be.cytomine.appengine.handlers.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 
 import java.io.File;
@@ -13,10 +15,41 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Comparator;
 
-public class FileSystemStorageHandler implements FileStorageHandler {
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class FileSystemStorageHandler implements StorageHandler {
 
     @Value("${storage.base-path}")
     private String basePath;
+
+    public void saveStorageData(Storage storage , StorageData storageData) throws FileStorageException {
+        if (storageData.peek() == null) return;
+        while (!storageData.isEmpty()) {
+            StorageDataEntry current = storageData.poll();
+            String filename = current.getName();
+            String storageId = storage.getIdStorage();
+            // process the node here
+            if(current.getStorageDataType() == StorageDataType.FILE){
+                try {
+                    Path filePath = Paths.get(basePath, storageId, filename);
+                    Files.createDirectories(filePath.getParent());
+                    Files.write(filePath, current.getData());
+                } catch (IOException e) {
+                    String error = "Failed to create file " + filename;
+                    error += " in storage " + storageId + ": " + e.getMessage();
+                    throw new FileStorageException(error);
+                }
+            }
+
+            if(current.getStorageDataType() == StorageDataType.DIRECTORY){
+                Storage modifiedStorage = new Storage(storageId + current.getName());
+                createStorage(modifiedStorage);
+
+            }
+
+        }
+    }
 
     @Override
     public void createStorage(Storage storage) throws FileStorageException {
@@ -48,21 +81,6 @@ public class FileSystemStorageHandler implements FileStorageHandler {
     }
 
     @Override
-    public void createFile(Storage storage, FileData file) throws FileStorageException {
-        String filename = file.getFileName();
-        String storageId = storage.getIdStorage();
-
-        try {
-            Path filePath = Paths.get(basePath, storageId, filename);
-            Files.write(filePath, file.getFileData());
-        } catch (IOException e) {
-            String error = "Failed to create file " + filename;
-            error += " in storage " + storageId + ": " + e.getMessage();
-            throw new FileStorageException(error);
-        }
-    }
-
-    @Override
     public boolean checkStorageExists(Storage storage) throws FileStorageException {
         return Files.exists(Paths.get(basePath, storage.getIdStorage()));
     }
@@ -73,27 +91,47 @@ public class FileSystemStorageHandler implements FileStorageHandler {
     }
 
     @Override
-    public void deleteFile(FileData file) throws FileStorageException {
-        String filename = file.getFileName();
-
-        try {
-            Path filePath = Paths.get(basePath, file.getStorageId(), filename);
-            Files.deleteIfExists(filePath);
-        } catch (IOException e) {
-            throw new FileStorageException("Failed to delete file " + filename);
+    public void deleteStorageData(StorageData storageData) throws FileStorageException {
+        String fileOrDirName = storageData.peek().getName();
+        if (storageData.peek().getStorageDataType() == StorageDataType.FILE) {
+            try {
+                Path filePath = Paths.get(basePath, storageData.peek().getStorageId(), fileOrDirName);
+                Files.deleteIfExists(filePath);
+            } catch (IOException e) {
+                throw new FileStorageException("Failed to delete file " + fileOrDirName);
+            }
         }
+        if (storageData.peek().getStorageDataType() == StorageDataType.DIRECTORY) {
+            Storage storage = new Storage(fileOrDirName);
+            deleteStorage(storage);
+        }
+
     }
 
     @Override
-    public FileData readFile(FileData emptyFile) throws FileStorageException {
-        String filename = emptyFile.getFileName();
-
+    public StorageData readStorageData(StorageData emptyFile) throws FileStorageException {
+        StorageDataEntry current = emptyFile.poll();
+        String filename = current.getName();
+        Path filePath = Paths.get(basePath, current.getStorageId(), filename);
         try {
-            Path filePath = Paths.get(basePath, emptyFile.getStorageId(), filename);
-            byte[] data = Files.readAllBytes(filePath);
-            emptyFile.setFileData(data);
+            Files.walk(filePath).forEach(path -> {
+                if (Files.isRegularFile(path)) {
+                    try {
+                        current.setData(Files.readAllBytes(path));
+                        current.setStorageDataType(StorageDataType.FILE);
+                        emptyFile.add(current);
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                }
+                if (Files.isDirectory(path)) {
+                    current.setStorageDataType(StorageDataType.DIRECTORY);
+                    emptyFile.add(current);
+                }
+            });
             return emptyFile;
-        } catch (IOException e) {
+        } catch (IOException | RuntimeException e) {
+            emptyFile.getQueue().clear();
             throw new FileStorageException("Failed to read file " + filename);
         }
     }
